+++ changes.md
@@ -9,6 +9,7 @@
   - simple access to branch data with "dump" draw option
 2. Elementary TTree::Draw
   - simple 1D/2D histograms
+  - configurable histogram like "px:py>>hist(50,-5,5,50,-5,5)"
 3. Full support of Float16_t and Double32_t types 
 4. Drawing of RooPlot objects, I/O support for RooFit classes